refactor(event-detail): split the two heavyweight tabs out of tabs.tsx
event-detail/tabs.tsx bundled all 7 workflow tab components into one file,
pulling 24 imports (13 panels) — the worst high-coupling finding in the repo
(threshold 16). Each tab only renders 1-6 panels, so the file's dependency set
was the union of all tabs' needs.

Extract the shared prop contract (`EventTabContentProps`) into
`tab-content-props.ts`, and the two heavyweight, multi-panel tabs (Overview:
6 panels; Operations: Cloudscape primitives + panels + most of the JSX) into
their own modules (`OverviewTab.tsx` / `OperationsTab.tsx`). The five remaining
tabs are thin single/double-panel wrappers that stay co-located. `tabs.tsx`
re-exports the extracted tabs so the only consumer (EventDetail.tsx) is
unchanged.

tabs.tsx now imports 8 modules (was 24) and is off the high-coupling list; each
new module depends only on the panels it renders (SOLID — single responsibility,
minimal coupling).

Relates #1418.

## Regression analysis
- Pure structural move: tab component bodies are byte-identical, only relocated;
  the barrel preserves the `./event-detail/tabs` public surface. typecheck passes
  and the EventDetail tab/wizard tests (10) pass unchanged. No behavior change.

## Physical impact
- NO-OP on CFn / deployed artifacts. Frontend source reorganization only.
